Time series analysis of apricot GDP values in Turkey and 10 year-forecasting
In this time series analysis project, we try to provide key insights into the gross production value of apricots in Turkey. While determining our data and its features such as long-term increases, non-stationarity, and diagnostic checks we realized the outliers that occurred between 2005 and 2010 may be the results of the politically driven effects or natural effects on agriculture that happened in Turkey in those years. We found proper fits for our high statistical forecast models such as ETS, TBATS, NNAR, and PROPHET. Among these assumptions checking the models and the outputs from their accuracy criteria, we reached our best model as the TBATS model. Furthermore, interpreting the forecasts and predictions for model graphs, considering our primary question ‘How has the gross production value of apricots in Turkey changed over time? We reached the 4-year forecast of the gross production value of apricots in Turkey. After reaching the conclusion that it was performing an increasing behavior, we managed to complete our analysis.